CLDIDE-2405; do not remove temp user if he is still present at another WS;

+++ codenvy-hosted-workspace-events/src/main/java/com/codenvy/workspace/listener/WorkspaceRemovalListener.java
@@ -82,7 +82,7 @@ public class WorkspaceRemovalListener implements RemovalListener<String, Boolean
 
                     for (Member member : members) {
                         final Map<String, String> preferences = preferenceDao.getPreferences(member.getUserId());
-                        if (parseBoolean(preferences.get("temporary"))) {
+                        if (parseBoolean(preferences.get("temporary")) && memberDao.getUserRelationships(member.getUserId()).isEmpty()) {
                             userDao.remove(member.getUserId());
                         }
                     }
